Announcements app deletd and merged into course app
