+++ course/admin.py
@@ -1,6 +1,7 @@
 from django.contrib import admin
 
 from .models import (
+    Announcement,
     Chapter,
     Course,
     CourseHistory,
@@ -123,6 +124,18 @@ class PageAdmin(admin.ModelAdmin):
     )
 
 
+class AnnouncementAdmin(admin.ModelAdmin):
+    list_display = (
+        "id",
+        "course",
+        "body",
+        "is_pinned",
+        "created_on",
+        "modified_on",
+    )
+    search_fields = ("body",)
+
+
 admin.site.register(Course, CourseAdmin)
 admin.site.register(CourseHistory, CourseHistoryAdmin)
 admin.site.register(Chapter, ChapterAdmin)
@@ -130,3 +143,4 @@ admin.site.register(Section, SectionAdmin)
 admin.site.register(Notification, NotificationAdmin)
 admin.site.register(Schedule, ScheduleAdmin)
 admin.site.register(Page, PageAdmin)
+admin.site.register(Announcement, AnnouncementAdmin)
