+++ course/models.py
@@ -161,3 +161,14 @@ class Page(models.Model):
 
     def __str__(self):
         return self.title
+
+
+class Announcement(models.Model):
+    course = models.ForeignKey(Course, on_delete=models.CASCADE)
+    body = models.TextField()
+    is_pinned = models.BooleanField(default=False)
+    created_on = models.DateTimeField(auto_now_add=True)
+    modified_on = models.DateTimeField(auto_now=True)
+
+    def __str__(self):
+        return self.body
